requirements.txt
Flask>=2.0
pymongo>=4.0
python-dotenv>=1.0
google-analytics-data>=0.11.0
redis==5.0.1
rq==1.15.1
gunicorn
